Add support to not eable eithe heating or cooling if temp difference is 2 degree

+++ src/main/java/org/conserve/Rooms.java
@@ -15,7 +15,7 @@ abstract class Rooms {
         this.roomID = roomID;
     }
 
-    public double getTemperature(){
+    public float getTemperature(){
         return temperature;
     }
